+++ lib/rorvswild/plugin/action_mailer.rb
@@ -3,8 +3,8 @@ module RorVsWild
     class ActionMailer
       def self.setup
         return if @installed
-        return unless defined?(::ActiveSupport::Notifications.subscribe)
-        ::ActiveSupport::Notifications.subscribe("deliver.action_mailer", new)
+        return unless defined?(ActiveSupport::Notifications.subscribe)
+        ActiveSupport::Notifications.subscribe("deliver.action_mailer", new)
         @installed = true
       end