+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Instrument ActiveSupport cache
+* Instrument Rails.cache via ActiveSupport notifications
 * Read deployment's revision from $KAMAL_VERSION
 
 ## 1.9.1 (2025-02-20)